deep-learning/requirements.txt
neo4j
numpy
scipy
python-dotenv
matplotlib
networkx
scikit-learn